MH-13671 OAI-PMH autorepublish should actively query the mp from am
The OAI-PMH publication service and download distribution service are often in error state if you run some fix-metadata-and-republish-to-oaipmh workflows. In case of error you will see an error message like: can not distribute catalog, with inner message: file not found. The mediapackage, the oaipmh autorepublish code get from the asset manager (through ActiveMQ) contains URLs pointing to the working file repository. But in some situations, the workflow instance is finished and the associated files are removed. This is the root cause of the problem.

This patch query the asset manager for the snapshot and uses the media package from the snapshot for further processing.